make parameter for tuning maximum depth of tree

make parameter for tuning maximum depth of tree

Usage

maxdepth_par(maxdepth_min = 2, maxdepth_max = 12)

Arguments

maxdepth_min

An integer specifying the minimum value in the parameter grid

maxdepth_max

An integer specifying the maximum value in the parameter grid

Value

A quantitative parameter, from {dials} for maxdepth

Examples

dials::grid_max_entropy(
   maxdepth_par(maxdepth_min = 2, maxdepth_max = 5),
   alpha_par(alpha_min = 0.10, alpha_max = 0.001),
   trim_par(trim_min = 0.1, trim_max = 0.3),
   size = 10
   )
#> # A tibble: 10 × 3
#>    maxdepth_par alpha_par trim_par
#>           <int>     <dbl>    <dbl>
#>  1            5    0.0842    0.122
#>  2            5    0.0456    0.289
#>  3            3    0.0185    0.281
#>  4            2    0.0369    0.230
#>  5            5    0.0796    0.213
#>  6            3    0.0637    0.102
#>  7            3    0.0757    0.260
#>  8            5    0.0181    0.215
#>  9            2    0.0133    0.114
#> 10            4    0.0992    0.182