make parameter for tuning level of amount of outlier trimmed in instability tests

make parameter for tuning level of amount of outlier trimmed in instability tests

Usage

trim_par(trim_min = 0.1, trim_max = 0.5)

Arguments

trim_min

An integer specifying the minimum value in the parameter grid

trim_max

An integer specifying the maximum value in the parameter grid

Value

A quantitative parameter, from {dials} for maxdepth

Examples

dials::grid_max_entropy(
   maxdepth_par(maxdepth_min = 2, maxdepth_max = 5),
   alpha_par(alpha_min = 0.10, alpha_max = 0.001),
   trim_par(trim_min = 0.1, trim_max = 0.3),
   size = 10
   )
#> # A tibble: 10 × 3
#>    maxdepth_par alpha_par trim_par
#>           <int>     <dbl>    <dbl>
#>  1            5   0.0118     0.165
#>  2            5   0.0856     0.279
#>  3            5   0.0574     0.120
#>  4            3   0.00844    0.210
#>  5            2   0.0992     0.172
#>  6            3   0.0530     0.145
#>  7            2   0.0480     0.255
#>  8            4   0.0981     0.273
#>  9            5   0.0192     0.269
#> 10            5   0.0978     0.187